refactor(inputs): 明确知识载体契约目录
