+++ src/cli/commands/rollback.ts
@@ -14,7 +14,7 @@ import {
   managedRecordId,
   resolveManagedDir,
 } from '../../managed/index.js';
-import type { ArtifactKind } from '../../artifacts/contracts.js';
+import type { ArtifactKind } from '../../knowledge-artifacts/contracts.js';
 import type { ManagedDecision } from '../../managed/contracts.js';
 
 const SUPPORTED_KINDS: ArtifactKind[] = ['skill'];
