+++ src/doctor/index.ts
@@ -13,7 +13,7 @@ import { basename, dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { runFileSuffix } from '../measurement-artifacts/file-names.js';
 import { discoverVariants, resolveArtifacts } from '../inputs/skill-loader.js';
-import type { Artifact } from '../artifacts/contracts.js';
+import type { Artifact } from '../knowledge-artifacts/contracts.js';
 import type {
   DoctorContext,
   DoctorOutcome,
