+++ src/inputs/eval-config.ts
@@ -2,7 +2,7 @@ import { readFileSync, existsSync } from 'node:fs';
 import { resolve, dirname, isAbsolute, join } from 'node:path';
 import { parseYaml } from './load-samples.js';
 import type { EvalBudget, EvalConfig, EvalConfigVariant } from './contracts/config.js';
-import type { ExperimentRole } from '../artifacts/contracts.js';
+import type { ExperimentRole } from '../knowledge-artifacts/contracts.js';
 import type { JudgeConfig } from '../grading/contracts/config.js';
 import type { RemoteGitRef, VariantSpec } from './contracts/variant.js';
