+++ src/knowledge-artifacts/contracts.ts
