+++ test/doctor/rules.test.ts
@@ -13,7 +13,7 @@ import {
   getRegisteredRules,
   __resetCustomRulesForTest,
 } from '../../src/doctor/rules.js';
-import type { Artifact } from '../../src/artifacts/contracts.js';
+import type { Artifact } from '../../src/knowledge-artifacts/contracts.js';
 import type { Sample } from '../../src/inputs/contracts/sample.js';
 import type { DoctorContext, DoctorRule } from '../../src/doctor/contracts.js';
